nurs controller - details

+++ DomainModel/Nurse.cs
@@ -6,9 +6,9 @@ using System.Threading.Tasks;
 
 namespace DomainModel
 {
-    public class Nurse : Person
-    {
-        public int SiretNumber { get; set; }
-        public virtual ICollection<Appointment> Appointments { get; set; }
-    }
+	public class Nurse : Person
+	{
+		public long SiretNumber { get; set; }
+		public virtual ICollection<Appointment> Appointments { get; set; }
+	}
 }
